requirements.txt
# Already have (verify versions)
ultralytics>=8.0.0
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0
fastapi>=0.104.0
uvicorn>=0.24.0

# New additions
# face-recognition>=1.3.0       # Face recognition (requires cmake + dlib) - Commented out due to build issues
deepface>=0.0.79              # Fallback face recognition (easier install)
faiss-cpu>=1.7.4              # Vector similarity search
sqlalchemy>=2.0.0             # DB layer
python-multipart>=0.0.6       # File upload support
websockets>=12.0              # WebSocket support
Pillow>=10.0.0                # Image processing
numpy>=1.24.0
scipy>=1.11.0                 # Cosine similarity
imageio>=2.31.0               # Video writing
imageio-ffmpeg>=0.4.8         # FFMPEG backend for imageio
pydantic>=2.0.0               # Request/response models
PyYAML>=6.0.0                 # Config loading for training and edge runtime
onnx>=1.15.0                  # ONNX export/validation
onnxruntime>=1.17.0           # CPU inference and quantization tooling
albumentations>=1.4.0         # Training-only blur/noise/weather augmentation support
